requirements.txt
Flask==1.1.1
Flask-WTF==1.0.0
google-cloud-logging==1.12.1
google-cloud-datastore==1.9.0
google-cloud-storage==1.16.1
google-cloud-tasks==2.7.1
